Fix: Move decorator to sub-directory of common
