+++ src/admin/v1/auth.controller.ts
@@ -3,7 +3,7 @@ import { Controller, Get, Query, UseGuards } from '@nestjs/common';
 
 // Custom Packages
 import { RoleGuard } from '@app/authorization/guard/roles.guard';
-import { AllowRole } from '@app/decorator/role/roles.decorator';
+import { AllowRole } from '@app/common/decorator/role/roles.decorator';
 import { AdminV1Service } from './auth.service';
 import { AdminV1ControllerDocs, AdminV1ListUserDocs } from './docs';
 import { AdminV1ListUserQuery } from './dto';
